// src/app/data/models/user.model.ts
export class UserModel {

    id: number;
    codigo: string;
    nombre: string;
    tipo_doc: string;
    numeroDoc: string;
    telefono: string;
    celular: string;
    email: string;
    direccion: string;
    username: string;
    password: string;
    created: number;
    modified: number;
    business_id: number;
    setting_id: number;

}